fix(ci): use github.action_path for with-retry.sh in composites; clean up test placement
Critical: connector-integration-tests and setup-deps used ${GITHUB_WORKSPACE}
to locate with-retry.sh. When consumed cross-repo (e.g. from a connector), the
inner actions/checkout overwrites GITHUB_WORKSPACE with the caller's tree where
the script doesn't exist. Switch both actions to
${{ github.action_path }}/../../scripts/with-retry.sh, which always resolves
relative to the SDK action's own directory.

Test cleanup:
- TestSslContextPublicCertificates moved from unit to integration test file; the
  has_internet_connection() guard dropped so the network call is the assertion
  (guard was silently skipping under --disable-socket). Unused imports (socket,
  aiohttp, httpx) removed from unit file.
- test_record_activity_updates_last_activity moved back to unit tests: it only
  mutates WorkerHealthServer state and never starts a server, so using a fixed
  port=0 makes _find_free_port() unnecessary and keeps it out of CI integration.
